refactor: character component
