+++ apps/Cybar/components/Characters/character.tsx
