+++ apps/Cybar/components/Characters/characterIds.type.ts
